pyproject.toml
[tool.poetry]
name = "library_system"
version = "0.1.0"
description = "simple library system"
authors = ["Bishnu Bhattarai <bishnusyangja@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
django-extensions = "^3.2.3"
psycopg2 = "^2.9.9"
djangorestframework-jwt = "^1.11.0"
django-cors-headers = "^4.3.1"
model-bakery = "^1.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
